🚚 organize library
