+++ packages/treetrunks/src/flat.ts
@@ -1,0 +1,3 @@
+export type Flat<R extends { [K in PropertyKey]: any }> = {
+	[K in keyof R]: R[K]
+}
